:wrench: move DrealList within hydrated pages
