+++ src/controllers/dreal/getDrealPage.ts
@@ -1,9 +1,9 @@
 import { userRepo } from '@dataAccess'
 import asyncHandler from '../helpers/asyncHandler'
 import routes from '../../routes'
-import { DrealListPage } from '@views/legacy-pages'
 import { ensureRole } from '@config'
 import { v1Router } from '../v1Router'
+import { DrealListPage } from '@views'
 
 v1Router.get(
   routes.ADMIN_DREAL_LIST,
