+++ src/views/index.ts
@@ -15,6 +15,7 @@ import {
   SignalerDemandeAbandon,
   SignalerDemandeRecours,
   Signup,
+  DrealList,
 } from './pages'
 
 export const StatistiquesPage = (props: Parameters<typeof Statistiques>[0]) =>
@@ -134,4 +135,12 @@ export const SignupPage = (props: Parameters<typeof Signup>[0]) =>
     Component: Signup,
     props,
     hydrate: false,
+  })
+
+export const DrealListPage = (props: Parameters<typeof DrealList>[0]) =>
+  makeHtml({
+    Component: DrealList,
+    props,
+    hydrate: true,
+    pageName: 'drealList', // This must match the Component name
   })
