+++ src/views/legacy-pages/index.ts
@@ -5,7 +5,6 @@ import AdminAppelOffre from './adminAppelOffre'
 import AdminNotifyCandidates from './adminNotifyCandidates'
 import AdminRegenerateCertificates from './adminRegenerateCertificates'
 import AdminUsers from './adminUsers'
-import DrealList from './drealList'
 import ImportCandidates from './importCandidates'
 import InvitationList from './invitationList'
 import NotificationList from './notificationList'
@@ -19,7 +18,6 @@ const AdminNotifyCandidatesPage = makePresenterPage(AdminNotifyCandidates)
 const AdminRegenerateCertificatesPage = makePresenterPage(AdminRegenerateCertificates)
 const ImportCandidatesPage = makePresenterPage(ImportCandidates)
 const ListProjectsPage = makePresenterPage(ListProjects)
-const DrealListPage = makePresenterPage(DrealList)
 const InvitationListPage = makePresenterPage(InvitationList)
 const NotificationListPage = makePresenterPage(NotificationList)
 const SuccessOrErrorPage = makePresenterPage(SuccessOrError)
@@ -33,7 +31,6 @@ export {
   ImportCandidatesPage,
   ListProjectsPage,
   AdminNotifyCandidatesPage,
-  DrealListPage,
   InvitationListPage,
   NotificationListPage,
   SuccessOrErrorPage as SuccessPage,
