+++ src/views/pages/drealList/index.ts
@@ -1,0 +1,1 @@
+export * from './DrealList';
