+++ src/views/pages/index.ts
@@ -13,3 +13,4 @@ export * from './Error'
 export * from './Signup'
 export * from './StatistiquesPage'
 export * from './UploadLegacyModificationFilesPage'
+export * from './drealList'
